+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ascendant Sequence filter.
 - Descendant Sequence filter.
 - Extreme Sequence filter.
+- Frequent Value filter.
 - Opposite Pattern filter.
 - Pair filter.
 - Repeating Pattern filter.